requirements.txt
langchain
openai
faiss-cpu
streamlit
tqdm
python-dotenv
pypdf
